:recycle: move object to object_util.go
